+++ grunt/copy.js
@@ -4,20 +4,36 @@ module.exports = {
         files: [
             {
                 expand: true,
-                src: '*',
-                cwd: 'build/vendor/font-awesome/fonts/',
+                flatten: true,
+                src: [
+                    'fontawesome-webfont.eot',
+                    'fontawesome-webfont.svg',
+                    'fontawesome-webfont.ttf',
+                    'fontawesome-webfont.woff',
+                    'fontawesome-webfont.woff2'
+                ],
+                filter: 'isFile',
+                cwd: 'node_modules/font-awesome/fonts/',
                 dest: 'out/flow/src/fonts/'
             },
+            /* No need to copy FlexSlider fonts because replaced by FontAwesome
             {
                 expand: true,
                 src: '*',
                 cwd: 'build/vendor/jquery-flexslider2/fonts/',
                 dest: 'out/flow/src/fonts/'
             },
+            */
             {
                 expand: true,
-                src: '*',
-                cwd: 'build/vendor/photoswipe/img/',
+                flatten: true,
+                src: [
+                    'default-skin.png',
+                    'default-skin.svg',
+                    'preloader.gif'
+                ],
+                filter: 'isFile',
+                cwd: 'node_modules/photoswipe/dist/default-skin/',
                 dest: 'out/flow/src/bg/'
             }
         ]