+++ grunt/uglify.js
@@ -3,10 +3,10 @@ module.exports = {
     theme: {
         files: {
             "out/flow/src/js/scripts.min.js": [
-                "build/vendor/bootstrap/js/bootstrap.js",
-                "build/vendor/bootstrap-select/js/bootstrap-select.js",
+                "node_modules/bootstrap/dist/js/bootstrap.js",
+                "node_modules/bootstrap-select/dist/js/bootstrap-select.js",
                 "build/vendor/jquery-bootstrap-validation/js/jqBootstrapValidation.js",
-                "build/vendor/jquery-unveil/js/jquery.unveil.js",
+                "node_modules/src/jquery.unveil2.js",
                 "build/js/main.js"
             ],
             "out/flow/src/js/pages/compare.min.js": "build/js/pages/compare.js",
@@ -18,14 +18,27 @@ module.exports = {
 
     vendor: {
         files: {
-            "out/flow/src/js/libs/jquery.min.js":                "build/vendor/jquery/js/jquery.js",
+            "out/flow/src/js/libs/jquery.min.js":                "node_modules/jquery/dist/jquery.js",
             "out/flow/src/js/libs/jquery.cookie.min.js":         "build/vendor/jquery-cookie/js/jquery.cookie.js",
             "out/flow/src/js/libs/jquery.flexslider.min.js":     "build/vendor/jquery-flexslider2/js/jquery.flexslider.js",
             "out/flow/src/js/libs/jquery-ui.min.js":             "build/vendor/jquery-ui/js/jquery-ui.js",
             "out/flow/src/js/libs/jqBootstrapValidation.min.js": "build/vendor/jquery-bootstrap-validation/js/jqBootstrapValidation.js",
-            "out/flow/src/js/libs/photoswipe.min.js":            "build/vendor/photoswipe/js/photoswipe.js",
-            "out/flow/src/js/libs/photoswipe-ui-default.min.js": "build/vendor/photoswipe/js/photoswipe-ui-default.js"
+            "out/flow/src/js/libs/photoswipe.min.js":            "node_modules/photoswipe/dist/photoswipe.js",
+            "out/flow/src/js/libs/photoswipe-ui-default.min.js": "node_modules/photoswipe/dist/photoswipe-ui-default.js"
         }
+    },
+
+    vendorBootstrapSelectI18N: {
+        files: [
+            {
+                expand: true,
+                src:    ['*.js', '!*.min.js'],
+                cwd:    "node_modules/bootstrap-select/dist/js/i18n/",
+                dest:   "out/flow/src/js/libs/bootstrap-select-i18n/",
+                ext:    ".min.js",
+                extDot: "last"
+            }
+        ]
     },
 
     widgets: {